Move Particle3DContainer class to new file
